test.c: Doesn't really "fix" anything but yeah.
This and the previous commit are just a result of fixing bad copy/paste
errors.  This technically didn't result in broken behavior because when
the process quits, it basically frees all its resources anyway but it
might not always work well in the future if stuff may involve like
sockets or something.