Add plans for a music sequence format.
Something which mostly matches the capabilities of the synth engine so
it can be made use of.  Considered MOD but it has too many particular
considerations for the platform on which it was made for.  I could've
used an existing engine but that would be another dependency and
wouldn't really show off the usefulness of the synth engine.  There's no
reason an existing engine couldn't be used though and just mixed in as a
pair of stereo channels mixed in.