logging: Make logging callbacks just take the final string.
Hugely simplify _everything_ involving logging from the user's point of
view and adapting it for python requiring no more python-specific
shenanigans, making the low-level API identical between C and python.
Now the logging callback will be called by an intermediary which will
handle all the printf-style arguments, and just pass on the final string
on to the callback.
Also included a header which has been missing from the repo for well
over a week.