synth.c: Tons of rewriting, should work mostly the same.
Mostly just got rid of a _lot_ of repetitive, error-prone code, made
buffers "opaque" to the rest of the API to smooth over the difference
between channel buffers and memory buffers.  Rewrite much of the synth
logic to be simpler and also somewhat less repetitive where it makes
sense.  Major differences are that PINGPONG mode is gone because it's
dumb and one can just get a similar effect by many other means, as well
as source buffers no longer inherently loop, but will return when the
buffer is exhausted.  This can also be worked around.