audio.py: Bunch of crap.
Audio sequencer can now be tuned and has the ability to eventually be
fed values from the speed field and return speeds based on absolute
speed, frequency or notes.
Added the ability to provide callback functions for handling data when
read from a file.  Thought I could use it, didn't end up needing it,
will keep it and fix it if it's broken.
All of this is untested.