audio.py: Try to get envelopes working.
This has required a lot of changes all around, and it still doesn't work
right.  Event frequently don't trigger.  Much of what had to be done
involved allowing the script to be able to only return as much buffer as
it can at once, and allow the engine to increase latency as-needed, so
the sequence can run more decoupled from the sample rate and be
processed in millisecond-sized blocks.