synth.c: Start of setting up buffers and things for filter.
Hopefully all the internal buffer filling stuff is set up properly at
this point.  I was a bit lazy about the internal buffer not being a
proper ring buffer but I might try for that, who knows.  Avoiding
unnecessaril memcpy would be nice but it should only be a real problem
on large, slow filters anyway where the filter will be most of the time
by quite a bit.