More farting around.
sequencer is a bit more strict about extra fields.  The rationale in the
past was that extra fields could be space for comments, but now there's
support for actual comments that the sequencer will never see so i don't
need to be able to support that confusion.  Makes catching mistakes
easier.