More screwing around.
Trying to establish some common macros to use in making the sequences
easier to write and read.
Added more debugging information to help in diagnosing problems with
sequences but also otherwise in helping me diagnose problems with the
sequencer itself.
Removed an unnecessary memory allocation.
Fixed a bug in filter allocation.