I guess convert to heap modules/types.
Seems tobe the way things are going or at least the future recommended
way to do things, and seems to allow more user flexibility so may as
well.  Also, try importing SDL and ctypes and getting the SDL_* pointer
types needed from ctypes and hopefully holding a reference to the types
will mean they won't be freed and when SDL users use those types, it'll
be the same type and type checking will be possible.