crustygamemodule: New tests pass now.
Had to do some real janky stuff to make it work but amazingly it does,
and it should be the most robust way for this sort of thing to work.  32
bit platform support might not work but if it is broken and there's
interest, i can probably make that work.  I don't have a 32 bit system
to test on though.