Start to torture test the sequencer and synth.
For now, it's just the beginning of Nutshell but there's a lot of crispy
noise that needs to be figured out.  No idea what's going on,
unfortunately.  Then finish the song and all parts to see how
performance goes.  So far on my system, the sequencer and synth running
on this song only uses ~1.5% CPU so it's much better than expected so
far.