audio.py: Added part-files and max-request tags.
part-files is untested, but in theory it should load multiple sequence
files atomically so when playback starts, they'll be in lockstep with
each other.  This isn't part of the engine though, just part of the
example.

max-request instructs the audio system not to request more than some
amount of milliseconds at a time so intermediate buffers aren't overrun
because the system latency/buffersize got too big, which would break
sequences in the past, if this happened.