sequencer.py: Made some errors more useful.
audio.py: Removed an extra line of useless information.
audio.py: Made an exception type more specific?

Overall, made modularizing some useful components easier.  For now it's
maybe not best to listen to the output as I don't entirely know what's
wrong but it'll produce a very loud rumble that might not be good for
equipment.  Some stuff might still be wrong, but it's a start to things
being more useful.