synth.c: Fix potential bad memory access.
In trying to optimize things, a problem was discovered, which actually
makes the code a bit slower, and a use of a branchless method didn't
gain any speed, and was actually slower.
The problem was that only adding or subtracting the loop length was used
once in case of overflows, but small loop length and high speed values
could cause invalid accesses, now fmod clamps the value as it should.