synth: Add ability to access buffers.
Also pass through the python buffer interface to these buffers.  A bit
of an "unsafe" API, especially with the output buffers, but whichever.
This should allow for some useful things like audio visualization or
whatever else.  It's exposed through the python buffers API so these
should be compatible with ndarrays and things like that.  Untested,
might be broken, and there's no convenience function to create an
InternalBuffer object from a Buffer object (yet) but that should be
simple.