Some fixes around readable output buffers.
Now None can be passed in as a filename in the python module to indicate
not to open a .wav file and just provide only output buffers with no
file or device output.
Referenced output channels now need to be released as referencing an
output channel will lock the audio device to keep the buffers valid
while in use, and it will be unlocked once all output channels are
released.  This will have no effect if there's no device output.