tilemap: Started relative layer positioning.
Also allowed arbitrary SDL_Textures to be refered to by a Layer so they
could be part of relative positioning.

Currently untested, but at least the python demo still works (Cat demo
will need to be adapted to it but it's a small change).

Honestly probably super broken because i don't properly know
trig/geometry.