Fix block copies.
The fix is an optional define in the hopes that maybe it could work
without the fix in the future.  The fix consumes double the video memory
for each tilemap, however the extra memory isn't used until the copy
function is called, so if it's never needed, it won't impact anything
else.