Test optional rendering.
Was an optimization I was considering that might be necessary at some
point but i'm not going to add the coplexity to the whole program.  it
didn't make much difference and even the most complex screen is using
less CPU time than I thought it was so it shouldn't be necessary. I'll
leave in the addition to the one screen for demonstration purposes,
though.