display.py: Allow tilemap to be smaller than view.
Also allow view port to be defined in pixels.  This required rewriting
how the viewport layer parameters are calculated and updated so there
might be regressions and it might just not work entirely right but it's
been tested somewhat to work.  mapedit hasn't been updated to work with
the new system.  Also updateregion() needs to be updated, and this has
been broken for some time anyway, just not in a way that prevented
things from functioning but now it probably is completely broken.