Big all around cleanup.
Moved a bunch of files around as things were getting very cluttered in
the repo root.  Adjusted all paths to work with the new locations.
Found a couple little bugs here and there in things, specifically the
scanning for references in tilemap.c was basically broken but now should
actually be working correctly.